Refactor diagram-related classes and move visibility helper to new Helpers class
