+++ src/DotUML.CLI/Diagram/Helpers.cs
@@ -1,7 +1,7 @@
-namespace DotUML.CLI.Models;
+namespace DotUML.CLI.Diagram;
 
 // Utility method to get visibility character
-public static class DiagramHelpers
+public static class Helpers
 {
     public static char GetVisibilityCharacter(string visibility) => visibility switch
     {
